+++ packages/electricsql_cli/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.3
+
+* Update default Electric service port
+* Better error handling
+
 ## 0.1.2
 
 * Update generation style